Feature: Search & Chore
The HTML for the cards has been removed to make it dynamic. All countries that have cases will now appear, instead of a fixed number. Removed the search page, and added a search bar to the main page. This will allow the user to search for any country that is displayed - all countries with cases.